# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Matrix",
    "description": "Agent-Driven Cyber Threat Simulator - AI-powered security testing platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/scans/": {
      "post": {
        "tags": [
          "Scans"
        ],
        "summary": "Create Scan",
        "description": "Create a new security scan.\n\nScans are queued for asynchronous execution via the RQ worker queue.\nFalls back to FastAPI BackgroundTasks if Redis is unavailable.",
        "operationId": "create_scan_api_scans__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Scans"
        ],
        "summary": "List Scans",
        "description": "List user's scans with pagination.",
        "operationId": "list_scans_api_scans__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Size"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ScanStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scans/{scan_id}": {
      "get": {
        "tags": [
          "Scans"
        ],
        "summary": "Get Scan",
        "description": "Get a specific scan by ID.",
        "operationId": "get_scan_api_scans__scan_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Scans"
        ],
        "summary": "Delete Scan",
        "description": "Delete a scan.",
        "operationId": "delete_scan_api_scans__scan_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scans/{scan_id}/start": {
      "post": {
        "tags": [
          "Scans"
        ],
        "summary": "Start Scan",
        "description": "Start a pending scan.",
        "operationId": "start_scan_api_scans__scan_id__start_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scans/{scan_id}/cancel": {
      "post": {
        "tags": [
          "Scans"
        ],
        "summary": "Cancel Scan",
        "description": "Cancel a running scan.",
        "operationId": "cancel_scan_api_scans__scan_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScanCreate": {
        "properties": {
          "target_url": {
            "type": "string",
            "title": "Target Url",
            "description": "URL of the target to scan"
          },
          "target_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Name",
            "description": "Friendly name for the target"
          },
          "scan_type": {
            "type": "string",
            "title": "Scan Type",
            "description": "Type of scan: full, quick, or custom",
            "default": "full"
          },
          "agents_enabled": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Agents Enabled",
            "description": "List of agent types to enable",
            "default": [
              "sql_injection",
              "xss",
              "auth",
              "api_security"
            ]
          },
          "enable_waf_evasion": {
            "type": "boolean",
            "title": "Enable Waf Evasion",
            "description": "Enable advanced WAF evasion techniques. WARNING: This may trigger security alerts on target systems. Only use for authorized penetration testing with explicit permission.",
            "default": false
          },
          "waf_evasion_consent": {
            "type": "boolean",
            "title": "Waf Evasion Consent",
            "description": "User acknowledges the risks of WAF evasion testing",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "target_url"
        ],
        "title": "ScanCreate",
        "description": "Schema for creating a new scan."
      },
      "ScanListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ScanResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "ScanListResponse",
        "description": "Schema for paginated scan list."
      },
      "ScanResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "target_url": {
            "type": "string",
            "title": "Target Url"
          },
          "target_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Name"
          },
          "scan_type": {
            "type": "string",
            "title": "Scan Type"
          },
          "status": {
            "$ref": "#/components/schemas/ScanStatus"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "total_vulnerabilities": {
            "type": "integer",
            "title": "Total Vulnerabilities",
            "default": 0
          },
          "critical_count": {
            "type": "integer",
            "title": "Critical Count",
            "default": 0
          },
          "high_count": {
            "type": "integer",
            "title": "High Count",
            "default": 0
          },
          "medium_count": {
            "type": "integer",
            "title": "Medium Count",
            "default": 0
          },
          "low_count": {
            "type": "integer",
            "title": "Low Count",
            "default": 0
          },
          "info_count": {
            "type": "integer",
            "title": "Info Count",
            "default": 0
          },
          "technology_stack": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Technology Stack",
            "default": []
          },
          "agents_enabled": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Agents Enabled",
            "default": []
          },
          "enable_waf_evasion": {
            "type": "boolean",
            "title": "Enable Waf Evasion",
            "default": false
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "target_url",
          "scan_type",
          "status",
          "progress",
          "created_at"
        ],
        "title": "ScanResponse",
        "description": "Schema for scan response."
      },
      "ScanStatus": {
        "type": "string",
        "enum": [
          "pending",
          "queued",
          "initializing",
          "running",
          "paused",
          "completed",
          "failed",
          "cancelled",
          "timed_out"
        ],
        "title": "ScanStatus",
        "description": "Scan status enumeration."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}